this code is WIP
It's broken, don't expect it to work.

Tests fail and new tests need to be written.

Defaults and behaviour will change.

I expect at least the following issues:

- seeking to a tell position is broken (and to a certain extent
  may remain broken)[1]
- a warning may be emitted more than once for the same error
- the same message might be emitted more than once in the same croak
  for a single error
- the readdelim code might infinite loop
- if no tranformation is needed (ie. badly encoded isn't permitted at all,
  and we fail or craok for any other thing we treat as an error), then
  we could avoid the accumulation buffer (future optimization?)
- the code might not follow the PerlIO API properly (which doesn't
  appear to be defined in detail)
- this code might make you double-facepalm

but I might be wrong.

[1] since errors are transformed into the replacement character we can't
use the position in the buffer to adjust the seek position, *but* that
might end up being our best guess.  The other issue with the accumulation
buffer should be fixable.