avoid a conceivable use after free
I don't believe this can be triggered now, but if a future change
calls hv_delete_common() on a stash with the key marked to be freed it
may cause a use after free in the future.

In the current codebase only the ENV_IS_CASELESS cases appear to end up
calling hv_delete_common() with HVkey_FREEKEY and that doesn't
operate on a stash.

I was a little worried that the free might be skipped if magic within
the intervening block threw, but the only object modified belong
to the implementation - ie. the AV backing @ISA's magic, so I don't
think it can be a problem.