IO: prevent long delay at end of t/io_sock.t under harness
harness waits for end of input reading from the stdout of the test
process, and had a ~15 second delay between reporting the final ok
from the io_sock.t test and reporting the final test result.

The cause is the code that claims to "Stop the server" was only
reconnecting to the server and then asking to terminate that
connection rather than terminating the connection itself.

Change the command to terminate the connection.

The code following that appeared to be trying to check that the server
had terminated, but had inverted logic, so it was testing that the server
*hadn't* terminated.  Simply fixing that would give us a flaky test
since there's no reason to believe the server would have terminated
in the brief time between the previous statement and the current one,
so re-work this to wait for the server and fail the test if it
isn't successful.

This change has the potential to block, but in that case the alarm()
would fire, and io_sock.t would fail on a signal and harness would
report the error, as desired.

We could also prevent the delay by reopening STDOUT to /dev/null, but
this would leave the child (server) process running for 15 seconds after
the parent io_sock.t process completed.