ask the compiler if it can find libquadmath and sanity check
This avoids having to ask the compiler (or using) its version
specific paths.

Also added sanity checks to ensure quadmath is quadmath.  This isn't
likely to be an issue for quadmath but is more likely for some of those
two or three letter named libraries.

It also ensures that the found quadmath is for the correct architecture.